Data structure for a grammar representation is implemented. Added ParserUnknownTokenException.
Grammar class has a table NonTerminal->Set<Rule> and a RuleExpression Evaluator that evaluates a RuleAction that is defined for each rule.
Rule is implemented as an array of Symbols, where Symbol can be neither Terminal or NonTerminal and with a RuleAction instance that can be evaluated using RuleExpressionEvaluator. Also, it returns a value that is the return value of the last expression in the sequence.
Terminal contains Token-object that is token that the Terminal is associated with.

ParserUnknownTokenException is defined for the special case, when the grammar-file-parser is not able to tokenize string, or for user-defined lexer, where the exception can be used for indicate such type of errors.